feat:new data integration

+++ libs/api/domains/icelandic-government-institutions/src/lib/models/invoiceGroups.model.ts
@@ -1,10 +1,10 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql'
+import { Field, Float, Int, ObjectType } from '@nestjs/graphql'
 import { PaginatedResponse } from '@island.is/nest/pagination'
 import { InvoiceGroup } from './invoiceGroup.model'
 
 @ObjectType('IcelandicGovernmentInstitutionsInvoiceGroups')
 export class InvoiceGroupCollection extends PaginatedResponse(InvoiceGroup) {
-  @Field(() => Int)
+  @Field(() => Float)
   totalPaymentsSum!: number
 
   @Field(() => Int)
